# MoveYourBum/MoveYourBum/OpenAPIs/swagger2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MoveYourBumAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Exercise/{id}": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseForView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Exercise"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseForView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Exercise"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Schedule/{id}": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleForView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Schedule"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleForView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Schedule"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/ScheduleExercise/{id}": {
      "get": {
        "tags": [
          "ScheduleExercise"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleExerciseForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleExerciseForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleExerciseForView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ScheduleExercise"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleExerciseForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleExerciseForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleExerciseForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleExerciseForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleExerciseForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleExerciseForView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ScheduleExercise"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ScheduleExerciseSet": {
      "get": {
        "tags": [
          "ScheduleExerciseSet"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleExerciseSetForView"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleExerciseSetForView"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleExerciseSetForView"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ScheduleExerciseSet"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleExerciseSetForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleExerciseSetForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleExerciseSetForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleExerciseSetForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleExerciseSetForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleExerciseSetForView"
                }
              }
            }
          }
        }
      }
    },
    "/api/ScheduleExerciseSet/{id}": {
      "get": {
        "tags": [
          "ScheduleExerciseSet"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleExerciseSetForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleExerciseSetForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleExerciseSetForView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ScheduleExerciseSet"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleExerciseSetForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleExerciseSetForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleExerciseSetForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleExerciseSetForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleExerciseSetForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleExerciseSetForView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ScheduleExerciseSet"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DayScheduleForView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "idDay": {
            "type": "integer",
            "format": "int32"
          },
          "dayDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "idSchedule": {
            "type": "integer",
            "format": "int32"
          },
          "scheduleName": {
            "type": "string",
            "nullable": true
          },
          "scheduleDescription": {
            "type": "string",
            "nullable": true
          },
          "scheduleExercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleExerciseForView"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ExerciseForView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "properTechniqueDescription": {
            "type": "string",
            "nullable": true
          },
          "muscleInvolvedDescription": {
            "type": "string",
            "nullable": true
          },
          "idExerciseType": {
            "type": "integer",
            "format": "int32"
          },
          "exerciseTypeName": {
            "type": "string",
            "nullable": true
          },
          "photos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExercisePhotoForView"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ExercisePhotoForView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "fileUrl": {
            "type": "string",
            "nullable": true
          },
          "idExercise": {
            "type": "integer",
            "format": "int32"
          },
          "exerciseName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ScheduleExerciseForView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "annotation": {
            "type": "string",
            "nullable": true
          },
          "idSchedule": {
            "type": "integer",
            "format": "int32"
          },
          "idExercise": {
            "type": "integer",
            "format": "int32"
          },
          "exerciseName": {
            "type": "string",
            "nullable": true
          },
          "sets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleExerciseSetForView"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ScheduleExerciseSetForView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "plannedReps": {
            "type": "string",
            "nullable": true
          },
          "actualReps": {
            "type": "string",
            "nullable": true
          },
          "weightUsed": {
            "type": "string",
            "nullable": true
          },
          "idScheduleExercise": {
            "type": "integer",
            "format": "int32"
          },
          "scheduleExerciseName": {
            "type": "string",
            "nullable": true
          },
          "idDaySchedule": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "dayScheduleName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ScheduleForView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "scheduleExercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleExerciseForView"
            },
            "nullable": true
          },
          "daySchedules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DayScheduleForView"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}